mvp/python/pyproject.toml
[project]
name = "device-to-cloud-mvp"
version = "0.1.0"
dependencies = ["pydantic>=2.7","pyyaml>=6.0"]

[tool.pytest.ini_options]
addopts = "-q"
